Fix audit manager bindings
